Refactored core and moved classes around.
