+++ owltorrent-client/src/main/java/edu/rice/owltorrent/client/HelloWorldClient.java
@@ -1,7 +1,6 @@
 package edu.rice.owltorrent.client;
 
 import edu.rice.owltorrent.common.entity.Torrent;
-import edu.rice.owltorrent.core.Peer;
 import edu.rice.owltorrent.core.serialization.TorrentParser;
 import edu.rice.owltorrent.network.HandShakeListener;
 import java.io.File;
