--- owltorrent-client/src/main/java/edu/rice/owltorrent/client/Temp.java
@@ -1,3 +1,0 @@
-package edu.rice.owltorrent.client;
-
-public class Temp {}
