+++ owltorrent-common/src/test/java/edu/rice/owltorrent/common/util/EncryptorTest.java
@@ -1,4 +1,4 @@
-package edu.rice.owltorrent.core.serialization;
+package edu.rice.owltorrent.common.util;
 
 import static org.junit.Assert.*;
