+++ owltorrent-network/src/main/java/edu/rice/owltorrent/network/PeerLocator.java
@@ -1,5 +1,6 @@
-package edu.rice.owltorrent.core;
+package edu.rice.owltorrent.network;
 
+import edu.rice.owltorrent.common.entity.Peer;
 import edu.rice.owltorrent.common.entity.Torrent;
 import java.util.List;
